Add DynamicObject model. Refactor various things.
DynamicObject is a Scala trait that lets you mix dynamic properties
together with static properties - it is heavily inspired from the
Python object model (__slots__ and __dict__). Its current biggest
drawback is that configuration is required in order to provide
uniform access to static properties.
The few objects already defined in the model have been refactored
to include the new trait.

Also, a few service interfaces have been refactored to make use of
the Box container, when failure was a likely result.